refactor: APIUtils.js 제거 및 api/index.js 와 통합
