+++ front/src/App.js
@@ -3,7 +3,7 @@ import { useRecoilState } from "recoil";
 import "./index.css";
 import Alert from "react-s-alert";
 import Routes from "./router/Routes";
-import { getCurrentUser } from "./utils/APIUtils";
+import { getCurrentUser } from "./api";
 import { ACCESS_TOKEN, USER_PROTOTYPE } from "./constants";
 import "react-s-alert/dist/s-alert-default.css";
 import "react-s-alert/dist/s-alert-css-effects/slide.css";
@@ -22,7 +22,7 @@ const App = () => {
     getCurrentUser()
       .then((response) => {
         setUser({
-          currentUser: response,
+          currentUser: response.data,
           authenticated: true,
         });
         setLoading(false);
